+++ build-logic/src/main/kotlin/org.jabref.gradle.base.repositories.gradle.kts
@@ -1,0 +1,10 @@
+repositories {
+    mavenCentral()
+    maven { url = uri("https://s01.oss.sonatype.org/content/repositories/snapshots/") }
+    maven { url = uri("https://oss.sonatype.org/content/repositories/snapshots") }
+    maven { url = uri("https://jitpack.io") }
+    maven { url = uri("https://oss.sonatype.org/content/groups/public") }
+
+    // Required for one.jpro.jproutils:tree-showing
+    maven { url = uri("https://sandec.jfrog.io/artifactory/repo") }
+}
